pyproject.toml
[project]
name = "resolution-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langfuse>=2.0.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "python-dotenv>=1.2.1",
    "langchain>=1.1.0",
    "langgraph-cli[inmem]>=0.4.7",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "langsmith>=0.4.47",
    "fastapi>=0.115.0",
]

[project.scripts]
dev = "langgraph_cli.__main__:main"
